updated assignment5 folder structure
