--- Java_Assignment 5-8/src/Assignment5/main/MainClass.java
@@ -1,15 +1,0 @@
-package Assignment5.main;
-
-import Assignment5.data.DefaultInitialization;
-import Assignment5.singleton.SingletonExample;
-
-public class MainClass {
-    public static void main(String[] args) {
-        DefaultInitialization obj = new DefaultInitialization();
-        obj.printMemberVariables();
-        obj.printLocalVariables(); // will not print anything
-
-        SingletonExample singleton = SingletonExample.initialize("Hello Singleton");
-        singleton.print();
-    }
-}
