+++ Java_Assignment 5-8/src/Pavithra/Assignment/main/MainClass.java
@@ -1,0 +1,19 @@
+package Pavithra.Assignment.main;
+
+
+
+import Pavithra.Assignment.data.DefaultInitialization;
+import Pavithra.Assignment.singleton.SingletonExample;
+
+public class MainClass {
+    public static void main(String[] args) {
+        // First class: DefaultInitializationAssignment
+        DefaultInitialization obj = new DefaultInitialization();
+        obj.printMemberVariables();
+        obj.printLocalVariables();
+
+        // Second class: SingletonExample
+        SingletonExample singleton = SingletonExample.initialize("Hello from Singleton!");
+        singleton.print();
+    }
+}
